requirements.txt
Flask==2.3.2
Pillow==9.4.0
numpy==1.24.3
matplotlib==3.7.1
requests==2.31.0